Get cell types from GO ids

Return cell type information for specific Gene Ontology identifiers

go2cell(go_ids, species = "Homo sapiens")

Arguments

go_ids
A character vector of Gene Ontology identifiers.

species
A character of a species, can either be Homo sapiens (default) or Mus musculus.

Value

A dataframe of 5 columns: The first two correspond to cell types, that is, their Wikidata Identifier and their name. The two following columns correspond to the Gene Ontology IDs given, and the last column corresponds to the cell type's marker that led to the result.

Details

This function takes in a character vector of Gene Ontology (GO) identifiers, queries Wikidata (https://www.wikidata.org/) for genes related to said identifiers, and crosses this data with a local RDF database of cell type-marker correspondence, returning, at last, a dataframe of cell type information related to the given GO identifiers.

See also

Examples

# Return cell types related to 'muscle contraction' and 'metabolism'
go2cell(c("GO:0006936", "GO:0008152"))

# Return cell types related to 'response to zinc ion'
go2cell("GO:0010043")
if (FALSE) {
# IDs should use ':' not '_'
go2cell("GO_0010043")
}